+++ client/CHANGELOG.md
@@ -31,6 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `input_selection::Error::{InvalidOutputCount, InvalidOutputCount}` variants;
 - `Error::InvalidBech32Hrp`;
 - `Client::bech32_hrp_matches()`;
+- Bech32 HRP validation for provided addresses;
 
 ### Changed